+++ test/webpack.mockExclude.ts
@@ -1,0 +1,1 @@
+throw new Error("This file must be excluded");